\name{LPI.KAIRA-package}
\alias{LPI.KAIRA-package}
\docType{package}
\title{Lag Profile Inversion for KAIRA}
\description{ Lag Profile Inversion with KAIRA klp files}
\details{
  \tabular{ll}{
    Package: \tab LPI.KAIRA \cr
    Version: \tab 0.1-3 \cr
    Date: \tab 2013-10-22\cr
    License: FreeBSD\tab \cr
    Depends: LPI.gdf\tab \cr
    Built: \tab \cr
  }
  Index:

  \code{\link{LPI.KAIRA}} Lag Profile Inversion for KAIRA

  \code{\link{readLPIdata.KAIRA}} Raw data input function for KAIRA

  \code{\link{LPIparam.default.KAIRA}} Default LPI.KAIRA parameters
  
  

}

\author{
Ilkka Virtanen (University of Oulu, Finland) \cr \email{ilkka.i.virtanen@oulu.fi}
}

\keyword{ package }