\name{LPIparam.default.KAIRA}

\title{LPIparam.default.KAIRA}

\alias{LPIparam.default.KAIRA}

\description{LPI default parameters for KAIRA}

\usage{LPIparam.default.KAIRA()}

\arguments{None}

\value{
  A named list of LPI default arguments. See \link{LPI.KAIRA} for details.
}

\author{Ilkka Virtanen (University of Oulu, Finland) \cr
  \email{ilkka.i.virtanen@oulu.fi}}
